pyproject.toml
[project]
name = "enhanced-crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "strawgate",email = "williamseaston@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "ruff (>=0.11.6,<0.12.0)",
    "GitPython",
    "async-dns",
    "flask-autoindex (>=0.6.6,<0.7.0)",
]

[tool.poetry]
packages = [
    { include = "enhanced_crawler", from = "src" },
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
coverage = "^7.8.0"
syrupy = "^4.9.1"

[project.scripts]
enhanced-crawler = "enhanced_crawler.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"